refactor color conversion again to increase performance (x2.7 faster)
